Add Specialists Module

+++ src/store/users.js
@@ -4,25 +4,40 @@ export default {
     namespaced: true,
     state: {
         users: [],
+        alert: {
+            message: null,
+            status: null,
+        }
     },
     getters: {
         getUsers: (state) => state.users,
+        getAlert: (state) => state.alert,
     },
     mutations: {
         setUsers(state, users) {
             state.users = users;
         },
+        setAlert(state, alert) {
+            state.alert = alert;
+        }
     },
     actions: {
-        async addUser({commit}, data, successCB, errorCB) {
-            console.log(data);
+        async addUser({commit}, data) {
             try {
                 await UsersApi.create(data);
                 const response = await UsersApi.getAll();
                 commit('setUsers', response.data);
-                successCB();
+                commit('setAlert', {message: 'User has been successfully added!', status: 'success'});
             } catch (e) {
-                errorCB(e);
+                commit('setAlert', {message: e.response.data.message, status: 'error'});
+                console.error(e);
+            }
+        },
+        async fetchUserById({commit}, id) {
+            try {
+                return await UsersApi.getById(id);
+            } catch (e) {
+                commit('setAlert', {message: e.response.data.message, status: 'error'});
                 console.error(e);
             }
         },
@@ -34,13 +49,26 @@ export default {
                 console.error(e);
             }
         },
-        async removeUser({commit}, id) {
+        async editUser({commit}, data) {
             try {
-                await UsersApi.delete(id);
+                await UsersApi.update(data.id, data.formData);
                 const response = await UsersApi.getAll();
-                return commit('setUsers', response.data);
+                commit('setUsers', response.data);
+                commit('setAlert', {message: 'User has been successfully edited!', status: 'success'});
             } catch (e) {
+                commit('setAlert', {message: e.response.data.message, status: 'error'});
                 console.error(e);
+            }
+        },
+        async removeUser({commit}, id) {
+            if (confirm('Are you sure you want to delete this user?')) {
+                try {
+                    await UsersApi.delete(id);
+                    const response = await UsersApi.getAll();
+                    return commit('setUsers', response.data);
+                } catch (e) {
+                    console.error(e);
+                }
             }
         }
     }
